python/0164 - Weather App/requirements.txt
https://github.com/hyperskill/hs-test-python/archive/release.tar.gz
Flask
pyppeteer
urllib3
requests
flask-sqlalchemy
nest_asyncio